Fix BuildSharedServiceManager not releasing memory during build
The `BuildSharedServiceManager` is used by `RemapJarTask` and
`RemapSourcesJarTask` to prevent having to load mappings and potentially
TinyRemapper multiple times.
And while it does seemingly have a reference counter, only the thing as a whole
is reference counted, not the individual services. As such, if a build contains
many projects on different MC versions, even if the mappings for one of those
versions will no longer be used during the rest of the build, the manager will
still hold onto them, slowly but surely consuming more and more memory as the
build progresses.

This commit works around the issue by creating a separate
`BuildSharedServiceManager` per project (unless multiProjectOptimisation is
enabled), such that once all tasks of one project have completed, that whole
manager can be released.
For a proper fix one would likely want to do reference counting on the
SharedService level and/or use SoftReferences so cached data can be disposed of
if memory becomes scarce.

This workaround isn't ideal in that:
- Gradle is not required to run all tasks of one project before it goes to the
next one; in the worst case it could run `remapJar` of all projects before
running any `remapSourcesJar` tasks, effectively voiding this workaround. In
practice it seems to behave quite well though.
- Projects that used to be able to share services will now need to initialize
them multiple times. Given our Loom fork is primarily mainly used to develop
multi-version-heavy, not multi-project-heavy projects, this seems like an
acceptable trade-off (e.g. none of our libs are able to make use of the
sharing to begin with).